dao postgres finalizado

+++ server/dao-postgres/goal-service.js
@@ -11,10 +11,13 @@ const findByMatch = async(id, userId) => {
 }
 
 const saveGoal = async(goal) => {
-  const result = await con.query('INSERT INTO goals SET ?', goal)
+  const result = await con.query('INSERT INTO goals(minute, type, player, "matchId", "userId", matchpart) '+
+  'VALUES($1,$2,$3,$4,$5,$6) RETURNING *',
+  goal.minute, goal.type, goal.player, goal.matchId, goal.userId, goal.matchpart)
+  return result.rows[0]
 }
-const deleteGoal = async(id) => {
-  const result = await con.query('DELETE FROM goals WHERE id = $1 RETURNING *', [id])
+const deleteGoal = async(id, userId) => {
+  const result = await con.query('DELETE FROM goals WHERE id = $1 AND "userId"=$2 RETURNING *', [id, userId])
   return result.rows[0].id
 }
